flask_app/requirements.txt
flask
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests
requests
